requirements.txt
click==8.0.1
dnspython==1.16.0
docopt==0.6.2
eventlet==0.31.1
firebase-admin==5.0.1
Flask==1.0.2
Flask-Cors==3.0.10
Flask-Login==0.4.1
Flask-Session==0.3.1
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
grpcio==1.39.0
httplib2==0.19.1
idna==3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
msgpack==1.0.2
numpy==1.21.1
opencv-python==4.5.3.56
packaging==21.0
pipreqs==0.4.10
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.22
uritemplate==3.0.1
urllib3==1.26.6
websocket-client==1.1.0
Werkzeug==0.14.1
wget==3.2
WTForms==2.3.3
yarg==0.1.9

gunicorn